+++ features/promo/promo_checkout_common/src/main/java/com/tokopedia/promocheckout/common/domain/model/event/Cart.kt
@@ -15,13 +15,13 @@ data class Cart(
         val cartItems: List<CartItem> = emptyList(),
         @SerializedName("cashback_amount")
         @Expose
-        val cashbackAmount: Int = 0,
+        val cashbackAmount: Long = 0L,
         @SerializedName("count")
         @Expose
         val count: Int = 0,
         @SerializedName("display_price")
         @Expose
-        val displayPrice: Int = 0,
+        val displayPrice: Double = 0.0,
         @SerializedName("grand_total")
         @Expose
         val grandTotal: Int = 0,
@@ -36,10 +36,10 @@ data class Cart(
         val promocode: String = "",
         @SerializedName("promocode_cashback")
         @Expose
-        val promocodeCashback: Int = 0,
+        val promocodeCashback: Long = 0L,
         @SerializedName("promocode_discount")
         @Expose
-        val promocodeDiscount: Int = 0,
+        val promocodeDiscount: Long = 0L,
         @SerializedName("promocode_failure_message")
         @Expose
         val promocodeFailureMessage: String = "",
@@ -54,7 +54,7 @@ data class Cart(
         val totalConvFee: Int = 0,
         @SerializedName("total_price")
         @Expose
-        val totalPrice: Int = 0,
+        val totalPrice: Double = 0.0,
         @SerializedName("user")
         @Expose
         val user: User = User()
